refactor(util): getCardType을 util로 분리
