+++ src/hooks/useCardNumbers.tsx
@@ -1,6 +1,7 @@
 import { useState } from "react";
 import { isNumber, isUnderMaxLength } from "../validation/validate";
 import { setErrorMessage } from "../utils/setErrorMessage";
+import { getCardType } from "../utils/getCardType";
 
 const CARD_NUMBER = {
   CARD_NUMBER_MAX_LENGTH: 4,
@@ -12,22 +13,6 @@ const CARD_NUMBER_ERROR = {
   NOT_NUMBERIC_ERROR: "숫자만 입력 가능합니다.",
 } as const;
 
-const CARD_TYPE = {
-  VISA: {
-    NAME: "visa",
-    START_NUMBER: 40,
-    END_NUMBER: 49,
-  },
-  MASTER: {
-    NAME: "master",
-    START_NUMBER: 51,
-    END_NUMBER: 55,
-  },
-  DEFAULT: {
-    NAME: "default",
-  },
-} as const;
-
 export default function useCardNumbers() {
   const [cardNumbers, setCardNumbers] = useState(["", "", "", ""]);
   const [cardNumbersError, setError] = useState(["", "", "", ""]);
@@ -35,7 +20,6 @@ export default function useCardNumbers() {
 
   const validateCardNumbers = (value: string, index: number) => {
     if (!isUnderMaxLength(value.length, CARD_NUMBER.CARD_NUMBER_MAX_LENGTH)) {
-      console.log("dfsdfsfkdjfldfjs");
       setErrorMessage(
         cardNumbersError,
         CARD_NUMBER_ERROR.INVALID_LENGTH_ERROR,
@@ -73,20 +57,4 @@ export default function useCardNumbers() {
   };
 
   return { cardNumbers, cardType, cardNumbersError, onCardNumberChange };
-}
-
-function getCardType(firstTwoDigits: number) {
-  if (
-    firstTwoDigits >= CARD_TYPE.VISA.START_NUMBER &&
-    firstTwoDigits <= CARD_TYPE.VISA.END_NUMBER
-  ) {
-    return CARD_TYPE.VISA.NAME;
-  } else if (
-    firstTwoDigits >= CARD_TYPE.MASTER.START_NUMBER &&
-    firstTwoDigits <= CARD_TYPE.MASTER.END_NUMBER
-  ) {
-    return CARD_TYPE.MASTER.NAME;
-  } else {
-    return CARD_TYPE.DEFAULT.NAME;
-  }
 }
